pyproject.toml
[project]
name = "restaurant-bot-guardrails-complaints"
version = "0.1.0"
description = "Restaurant bot with guardrails and complaints handling."
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
